+++ MedSync.Presentation.Web/Controllers/LabTestController.cs
@@ -5,23 +5,31 @@ using MedSync.Core.Application.ViewModels.Users;
 using MedSync.Core.Domain.Enums;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using MedSync.Core.Application.ViewModels.LabTests;
+using MedSync.Presentation.Web.Middelware;
 namespace MedSync.Presentation.Web.Controllers
 {
     public class LabTestController : Controller
     {
         private readonly ILabTestService _labTestService;
         private readonly IHttpContextAccessor _httpContext;
+        private readonly ValidateUserSession _validateUserSession;
 
 
-        public LabTestController(ILabTestService labTestService, IHttpContextAccessor contextAccessor)
+        public LabTestController(ILabTestService labTestService, IHttpContextAccessor contextAccessor, ValidateUserSession validateUserSession)
         {
             _labTestService = labTestService;
             _httpContext = contextAccessor;
+            _validateUserSession = validateUserSession;
         }
 
         public async Task<IActionResult> Index()
 
         {
+            if(!_validateUserSession.IsValidUser(UserType.Administrador))
+            {
+               return RedirectToRoute(new { controller = "Account", action = "Login" });
+            }
+         
             UserViewModel userLogedIn = _httpContext.HttpContext!.Session.Get<UserViewModel>("user")!;
 
             return View(await _labTestService.GetAllByDoctorOfficeAsync(userLogedIn.DoctorOfficeId));
@@ -29,6 +37,10 @@ namespace MedSync.Presentation.Web.Controllers
 
         public IActionResult Add()
         {
+            if (!_validateUserSession.IsValidUser(UserType.Administrador))
+            {
+                return RedirectToRoute(new { controller = "Account", action = "Login" });
+            }
 
             return View("SaveLabTest");
         }
@@ -36,6 +48,11 @@ namespace MedSync.Presentation.Web.Controllers
         [HttpPost]
         public async Task<IActionResult> Add(SaveLabTestViewModel saveLabTestViewModel)
         {
+            if (!_validateUserSession.IsValidUser(UserType.Administrador))
+            {
+                return RedirectToRoute(new { controller = "Account", action = "Login" });
+            }
+
             ModelState.Remove("DoctorOfficeId");
             if(!ModelState.IsValid)
             {  
@@ -49,7 +66,10 @@ namespace MedSync.Presentation.Web.Controllers
 
         public async Task<IActionResult> Edit(int id)
         {
-
+            if (!_validateUserSession.IsValidUser(UserType.Administrador))
+            {
+                return RedirectToRoute(new { controller = "Account", action = "Login" });
+            }
             SaveLabTestViewModel saveLabTestViewModel = await _labTestService.GetByIdSaveViewModel(id);
 
             return View("SaveLabTest", saveLabTestViewModel);
@@ -58,7 +78,10 @@ namespace MedSync.Presentation.Web.Controllers
         [HttpPost]
         public async Task<IActionResult> Edit(SaveLabTestViewModel saveLabTestViewModel)
         {
-          
+            if (!_validateUserSession.IsValidUser(UserType.Administrador))
+            {
+                return RedirectToRoute(new { controller = "Account", action = "Login" });
+            }
             if (!ModelState.IsValid)
             {
 
@@ -72,6 +95,10 @@ namespace MedSync.Presentation.Web.Controllers
 
         public async Task<IActionResult> Delete(int id)
         {
+            if (!_validateUserSession.IsValidUser(UserType.Administrador))
+            {
+                return RedirectToRoute(new { controller = "Account", action = "Login" });
+            }
             LabTestViewModel? userViewModel = await _labTestService.GetById(id);
 
             return View("DeleteLabTest", userViewModel);
@@ -80,6 +107,10 @@ namespace MedSync.Presentation.Web.Controllers
         [HttpPost]
         public async Task<IActionResult> DeletePost(int id)
         {
+            if (!_validateUserSession.IsValidUser(UserType.Administrador))
+            {
+                return RedirectToRoute(new { controller = "Account", action = "Login" });
+            }
             await _labTestService.Delete(id);
 
             return RedirectToAction("Index");
