+++ MedSync.Presentation.Web/Controllers/UserController.cs
@@ -4,23 +4,31 @@ using MedSync.Core.Application.Helpers;
 using MedSync.Core.Application.ViewModels.Users;
 using MedSync.Core.Domain.Enums;
 using Microsoft.AspNetCore.Mvc.Rendering;
+using MedSync.Presentation.Web.Middelware;
 namespace MedSync.Presentation.Web.Controllers
 {
     public class UserController : Controller
     {
         private readonly IUserService _userService;
         private readonly IHttpContextAccessor _httpContext;
+        private readonly ValidateUserSession _validateUserSession;
 
 
-        public UserController(IUserService userService, IHttpContextAccessor contextAccessor)
+        public UserController(IUserService userService, IHttpContextAccessor contextAccessor, ValidateUserSession validateUserSession)
         {
             _userService = userService;
             _httpContext = contextAccessor;
+            _validateUserSession = validateUserSession;
         }
 
         public async Task<IActionResult> Index()
 
         {
+            if (!_validateUserSession.IsValidUser(UserType.Administrador))
+            {
+                return RedirectToRoute(new { controller = "Account", action = "Login" });
+            }
+
             UserViewModel userLogedIn = _httpContext.HttpContext!.Session.Get<UserViewModel>("user")!;
 
             return View(await _userService.GetAllByDoctorOfficeAsync(userLogedIn.DoctorOfficeId));
@@ -28,6 +36,12 @@ namespace MedSync.Presentation.Web.Controllers
 
         public IActionResult Add()
         {
+            if (!_validateUserSession.IsValidUser(UserType.Administrador))
+            {
+                return RedirectToRoute(new { controller = "Account", action = "Login" });
+            }
+
+
             ViewBag.UserTypes = Enum.GetValues(typeof(UserType))
                           .Cast<UserType>()
                           .Select(e => new SelectListItem
@@ -43,6 +57,12 @@ namespace MedSync.Presentation.Web.Controllers
         [HttpPost]
         public async Task<IActionResult> Add(SaveUserViewModel saveUserViewModel)
         {
+            if (!_validateUserSession.IsValidUser(UserType.Administrador))
+            {
+                return RedirectToRoute(new { controller = "Account", action = "Login" });
+            }
+
+
             ModelState.Remove("DoctorOfficeName");
             if(!ModelState.IsValid)
             {
@@ -64,6 +84,11 @@ namespace MedSync.Presentation.Web.Controllers
 
         public async Task<IActionResult> Edit(int id)
         {
+            if (!_validateUserSession.IsValidUser(UserType.Administrador))
+            {
+                return RedirectToRoute(new { controller = "Account", action = "Login" });
+            }
+
 
             SaveUserViewModel saveUserViewModel = await _userService.GetByIdSaveViewModel(id);
 
@@ -83,6 +108,12 @@ namespace MedSync.Presentation.Web.Controllers
         [HttpPost]
         public async Task<IActionResult> Edit(SaveUserViewModel saveUserViewModel)
         {
+            if (!_validateUserSession.IsValidUser(UserType.Administrador))
+            {
+                return RedirectToRoute(new { controller = "Account", action = "Login" });
+            }
+
+
             ModelState.Remove("Password");
             ModelState.Remove("ConfirmPassword");
             ModelState.Remove("DoctorOfficeName");
@@ -109,6 +140,12 @@ namespace MedSync.Presentation.Web.Controllers
 
         public async Task<IActionResult> Delete(int id)
         {
+            if (!_validateUserSession.IsValidUser(UserType.Administrador))
+            {
+                return RedirectToRoute(new { controller = "Account", action = "Login" });
+            }
+
+
             UserViewModel? userViewModel = await _userService.GetById(id);
 
             return View("DeleteUser", userViewModel);
@@ -117,6 +154,12 @@ namespace MedSync.Presentation.Web.Controllers
         [HttpPost]
         public async Task<IActionResult> DeletePost(int id)
         {
+            if (!_validateUserSession.IsValidUser(UserType.Administrador))
+            {
+                return RedirectToRoute(new { controller = "Account", action = "Login" });
+            }
+
+
             await _userService.Delete(id);
 
             return RedirectToAction("Index");
